+++ src/components/wine-detail/review-overview/ReviewOverview.tsx
@@ -1,0 +1,3 @@
+export default function ReviewOverview() {
+  return <div>리뷰 오버뷰</div>;
+}
